Add Colander benchmark
    - add ObjectType type for object serialization
    - add 'Differential' serializer to handle
      transformation during serialization
    - handle callable int via 'Callable' type
    - compare as strings in test
    - add sequence schema for 'many' case

+++ subjects/rf.py
@@ -10,7 +10,7 @@ __name__ = 'Django REST Framework'
 
 
 class SubRF(rf_serializers.Serializer):
-    w = rf_serializers.FloatField()
+    w = rf_serializers.IntegerField()
     x = rf_serializers.SerializerMethodField()
     y = rf_serializers.CharField()
     z = rf_serializers.IntegerField()
